+++ contracts/Transaction.sol
@@ -11,11 +11,11 @@ contract Transaction is Killable {
     bytes32 public desc;
     address public insurer;
     uint public coverage;
-    uint8 public maxCoverage;
-    uint8 public premium;
+    uint public maxCoverage;
+    uint public premium;
 
 
-    function Transaction(address _counterparty, bytes32 _name, bytes32 _desc, uint8 _max_coverage, uint8 _prem) public {
+    function Transaction(address _counterparty, bytes32 _name, bytes32 _desc, uint _max_coverage, uint _prem) public {
         require(_counterparty!=0x0);
         counterParty = _counterparty;
         owner = msg.sender;
@@ -44,8 +44,8 @@ contract Transaction is Killable {
             bytes32 _desc, 
             uint _value,
             uint _coverage,
-            uint8 _maxCoverage,
-            uint8 _premium,
+            uint _maxCoverage,
+            uint _premium,
             address _counterParty,
             address _insurer) {
                 return (name, desc, this.balance, coverage, maxCoverage,premium,counterParty,insurer);
@@ -84,11 +84,11 @@ contract Transaction is Killable {
     }
 
 
-    function setMaxCoverage(uint8 _coverage) onlyOwner public  {
+    function setMaxCoverage(uint _coverage) onlyOwner public  {
         maxCoverage = _coverage;
     }
 
-    function setPremium(uint8 _premium) onlyOwner public {
+    function setPremium(uint _premium) onlyOwner public {
         require(_premium>0);
         premium = _premium;
     }
